+++ LuckyoneUI/Core/Installer.lua
@@ -726,28 +726,34 @@ local function BuildInstallerData()
 	if Private.isRetail then
 		pages[pageIndex] = function()
 			local f = installerFrame
-			f.SubTitle:SetText(L["WarpDeplete profile"])
-			f.Desc1:SetText(L["Please click the button below to apply Luckyones profile for WarpDeplete."])
+			f.SubTitle:SetText(L["Mythic+ Addons"])
+			f.Desc1:SetText(L["Choose between WarpDeplete and MPlusTimer."])
 			f.Desc2:SetText(format('|cff4beb2c%s', L["Recommended step. Should not be skipped."]))
 			f.Option1:Show()
 			f.Option1:SetScript('OnClick', function() Private:Setup_WarpDeplete(true) end)
 			f.Option1:SetText(L["Setup WarpDeplete"])
+			f.Option2:Show()
+			f.Option2:SetScript('OnClick', function() Private:Setup_MPlusTimer(true) end)
+			f.Option2:SetText(L["Setup MPlusTimer"])
 		end
-		stepTitles[pageIndex] = 'WarpDeplete'
+		stepTitles[pageIndex] = L["Mythic+ Addons"]
 		pageIndex = pageIndex + 1
 
 		pages[pageIndex] = function()
 			local f = installerFrame
 			f.SubTitle:SetText(L["Cooldown Manager profile"])
-			f.Desc1:SetText(L["Choose between BetterCooldownManager and AyijeCDM."])
+			f.Desc1:SetText(L["Choose between SkironCooldownManager, AyijeCDM and BetterCooldownManager."])
 			f.Desc2:SetText(format('|cff4beb2c%s', L["Recommended step. Should not be skipped."]))
-			f.Desc3:SetText(L["For position adjustments use /bcdm or /acdm X and Y offset options."])
+			f.Desc3:SetText(L["For position adjustments use /scm /acdm /bcdm X and Y offset options."])
 			f.Option1:Show()
-			f.Option1:SetScript('OnClick', function() Private:Setup_BCDM(true) end)
-			f.Option1:SetText('BetterCooldownManager')
+			f.Option1:SetScript('OnClick', function() Private:Setup_SCM(true) end)
+			f.Option1:SetText('SkironCooldownManager')
 			f.Option2:Show()
 			f.Option2:SetScript('OnClick', function() Private:Setup_ACDM(true) end)
 			f.Option2:SetText('AyijeCDM')
+			f.Option3:Show()
+			f.Option3:SetScript('OnClick', function() Private:Setup_BCDM(true) end)
+			f.Option3:SetText('BetterCooldownManager')
 		end
 		stepTitles[pageIndex] = L["Cooldown Manager"]
 		pageIndex = pageIndex + 1