+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-﻿## Version 4.11 [29-04-2026]
+﻿## Version 4.12 [06-05-2026]
+
+**All WoW Versions:**  
+- 1080p improvements (BigWigs and Details)  
+- Added optional profile for the Baganator addon  
+- Updated frFR locales (Credits: Oniria)  
+- Updated graphics config section mouseover descriptions  
+
+**Retail:**  
+- Added basic edit mode specialization strings for all 40 specs (WIP)  
+- Added new Cooldown Manager section in /lucky config  
+- Added optional profile for the MPlusTimer addon  
+- Added SkironCooldownManager profile (Requires alpha)  
+- Adjusted default Y-offset of Boss- and Zone button slightly  
+
+## Version 4.11 [29-04-2026]
 
 **All WoW Versions:**  
 - ElvUI 15.13+ compatibility